Refactored stuff, added docu
